ACE_PRIME/HelloAF/src/requirements.txt
aio-pika==9.3.0
httpx==0.25.0
Jinja2==3.1.2
openai~=0.28.1
prompt-toolkit==3.0.39
psutil==5.9.5
pydantic-settings==2.0.3
Pygments==2.16.1
PyYAML==6.0.1
requests==2.31.0
Jinja2==3.1.2
openai~=0.28.1
python-dotenv~=1.0.0
